+++ IntegrationTester/IntegrationTester/SplitCardElementsViewController.swift
@@ -8,7 +8,6 @@
 import Foundation
 import UIKit
 import BasisTheoryElements
-import BasisTheory
 import Combine
 
 class SplitCardElementsViewController: UIViewController {