+++ IntegrationTester/UnitTests/TokenIntentTests.swift
@@ -1,4 +1,3 @@
-import BasisTheory
 import BasisTheoryElements
 import XCTest
 
@@ -6,7 +5,7 @@ final class TokenIntentTests: XCTestCase {
     private final var TIMEOUT_EXPECTATION = 5.0
 
     override func setUpWithError() throws {
-        BasisTheoryAPI.basePath = "https://api.flock-dev.com"
+        BasisTheoryElements.basePath = "https://api.flock-dev.com"
     }
 
     override func tearDownWithError() throws {}