+++ Package.swift
@@ -16,7 +16,7 @@ let package = Package(
     ],
     dependencies: [
         // Dependencies declare other packages that this package depends on.
-         .package(url: "https://github.com/Basis-Theory/basistheory-swift", .exact("0.7.0")),
+        .package(url: "https://github.com/Flight-School/AnyCodable", from: "0.6.0"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -24,7 +24,7 @@ let package = Package(
         .target(
             name: "BasisTheoryElements",
             dependencies: [
-                .product(name: "BasisTheory", package: "basistheory-swift"),
+                .product(name: "AnyCodable", package: "AnyCodable"),
             ],
             path: "BasisTheoryElements/Sources"
         )